Attempt to idle SWCLK in low state
This was an experiment to have SWCLK idle in the low state instead of
the high state. This means that each cycle is composed of a rising edge
and then a falling edge. This ends up not being a very natual way to
clock the data out and makes things much more complicated because:
* The target device reads and writes from/to SWDIO during the rising
  edges.
* To satisfy setup/hold times, it makes the most sense for the host to
  write data to SWDIO during the previous falling edge.
* If the first clock cycle starts with a rising edge then the target
  will sample the first bit before the host has placed the first real
  bit on SWDIO (which it will do on the upcoming falling edge of this
  cycle). This means that the first read by the target code during unit
  tests becomes a dummy read. This shouldn't be an issue on real
  targets but it makes the unit tests a bit more clumsy.
* The last clock cycle will end with a falling edge which will have
  written out the last valid bit on SWDIO but the test target won't
  read it as it never sees the rising edge. This would mean that
  another dummy clock cycle is required at the end to complete the
  last write.

While switching te idle state is possible, it require more dummy cycles
to have things work whereas the current implementation just does the
right thing on each cycle. It is also the clock cycling pattern used by
ARM in their own SWD host code:
 https://github.com/ARMmbed/DAPLink/blob/master/source/daplink/cmsis-dap/SW_DP.c#L37